bin: Add `batch-args`
This is essentially a reimplementation of xargs in Racket, but it is not
POSIX-compliant and uses its own interface, mostly because the POSIX
standard is a little strange in a lot of the details, and
POSIX-compliance would be needlessly complicated to implement for no
real gain.

The main advantage of this implementation over xargs, aside from
portability, is the ability to add arguments that appear before or after
the inserted arguments. This is especially useful if the arguments to be
inserted each ought to be prefixed with a flag. Doing this with ordinary
xargs is unnecessarily difficult.